+++ CHANGELOG.md
@@ -6,12 +6,12 @@
 
 * [ObjectServer] Added support for `SyncUser.isAdmin()` (#4353).
 * [ObjectServer] Added support for changing passwords through `SyncUser.changePassword()` (#4423).
-* [ObjectServer] Added support for `SyncConfigration.Builder.waitForInitialRemoteData()` (#4270).
+* [ObjectServer] Added support for `SyncConfiguration.Builder.waitForInitialRemoteData()` (#4270).
 * Transient fields are now allowed in model classes, but are implicitly treated as having the `@Ignore` annotation (#4279).
 * Added `Realm.refresh()` and `DynamicRealm.refresh()` (#3476).
 * Added `Realm.getInstanceAsync()` and `DynamicRealm.getInstanceAsync()` (#2299).
 * Added `DynamicRealmObject#linkingObjects(String,String)` to support linking objects on `DynamicRealm` (#4492).
-* Changelisteners will now auto-expand variable names to be more descriptive when using Android Studio.
+* Change listeners will now auto-expand variable names to be more descriptive when using Android Studio.
 
 ### Bug Fixes